requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
google-generativeai==0.8.3
openai==1.51.0
pydantic==2.9.0
httpx==0.27.2
python-dotenv==1.0.1
